RDKB-16174:Move proprietary steps to cpc repo
Reason for change: Moving proprietary steps from generic sysint to cpc
repo.
Test Procedure: Build and Verify
Risks: Medium

+++ source/scripts/init/service.d/service_lan/dhcp_lease_sync.sh
@@ -30,12 +30,13 @@ fi
 DHCP_LEASE_FILE_ARM="/nvram/dnsmasq.leases"
 DHCP_LEASE_FILE_ATOM="/nvram/dnsmasq.leases"
 DHCP_LEASE_FILE_ATOM_TMP="/tmp/dnsmasq.leases"
-PEER_COMM_DAT="/etc/dropbear/elxrretyt.swr"
-PEER_COMM_ID="/tmp/elxrretyt-$$.swr"
-CONFIGPARAMGEN="/usr/bin/configparamgen"
+PEER_COMM_ID="/tmp/elxrretyt.swr"
+if [ ! -f /usr/bin/GetConfigFile ];then
+    echo "Error: GetConfigFile Not Found"
+    exit 127
+fi
 
-
-$CONFIGPARAMGEN jx $PEER_COMM_DAT $PEER_COMM_ID
+GetConfigFile $PEER_COMM_ID
 scp -i $PEER_COMM_ID $DHCP_LEASE_FILE_ARM root@$ATOM_IP:$DHCP_LEASE_FILE_ATOM_TMP  > /dev/null 2>&1
 rpcclient $ATOM_ARPING_IP "flock $DHCP_LEASE_FILE_ATOM -c \"cp $DHCP_LEASE_FILE_ATOM_TMP $DHCP_LEASE_FILE_ATOM\""
 rpcclient $ATOM_ARPING_IP "rm -f $DHCP_LEASE_FILE_ATOM_TMP"
